Draft updates for -06 specification
Extend the discussion of reflexive variants and their use;
includes update of the appendix on converting tables in the
style of [RFC3743].  Improve description of tagging and clarify
that it doesn't apply to sequences.  Specify that root zone uses
".".  Add an appendix with an Indic Syllable Structure example.
Extend count attribute to allow maximal counts.